Link user with task and category

+++ internal/app/usecase/delete_task.go
@@ -4,11 +4,17 @@ import (
 	"context"
 
 	"github.com/tiagompalte/golang-clean-arch-template/internal/app/repository"
+	pkgErrors "github.com/tiagompalte/golang-clean-arch-template/internal/pkg/errors"
 	"github.com/tiagompalte/golang-clean-arch-template/pkg/errors"
 	usecasePkg "github.com/tiagompalte/golang-clean-arch-template/pkg/usecase"
 )
 
-type DeleteTaskUseCase usecasePkg.UseCase[string, usecasePkg.Blank]
+type DeleteTaskUseCase usecasePkg.UseCase[DeleteTaskUseCaseInput, usecasePkg.Blank]
+
+type DeleteTaskUseCaseInput struct {
+	UUID   string
+	UserID uint32
+}
 
 type DeleteTaskUseCaseImpl struct {
 	taskRepository repository.TaskRepository
@@ -20,10 +26,14 @@ func NewDeleteTaskUseCaseImpl(taskRepository repository.TaskRepository) DeleteTa
 	}
 }
 
-func (u DeleteTaskUseCaseImpl) Execute(ctx context.Context, uuid string) (usecasePkg.Blank, error) {
-	task, err := u.taskRepository.FindByUUID(ctx, uuid)
+func (u DeleteTaskUseCaseImpl) Execute(ctx context.Context, input DeleteTaskUseCaseInput) (usecasePkg.Blank, error) {
+	task, err := u.taskRepository.FindByUUID(ctx, input.UUID)
 	if err != nil {
 		return usecasePkg.Blank{}, errors.Wrap(err)
+	}
+
+	if task.UserID != input.UserID {
+		return usecasePkg.Blank{}, errors.Wrap(pkgErrors.NewInvalidUserError())
 	}
 
 	err = u.taskRepository.Delete(ctx, task)
